LanguageAgentTreeSearch/programming/requirements.txt
openai>=1.0.0
jsonlines
astunparse
python-dotenv
numpy
matplotlib
# Add any other dependencies your project requires
